feat : 함수 분리 및 주석 추가
- 장바구니 관련 함수 분리
- 주석추가
