+++ src/component/assets/_cart.scss
@@ -1,12 +1,3 @@
-@mixin flexAll($fd, $jc, $ac, $gp) {
-  display: flex;
-  flex-direction: $fd;
-  flex-wrap: nowrap;
-  justify-content: $jc;
-  align-items: $ac;
-  gap: $gp;
-}
-
 // 컴포넌트 전체
 .cart-component {
   margin: 80px 0;
