+++ src/pages/cart/main.js
@@ -12,6 +12,12 @@ import {
   removeClass,
   toggleClass,
 } from "../../lib";
+import { countChange, updateSelectedCount } from "./countedItem";
+import { deleteItem, deleteSelected } from "./delete";
+import selectAll from "./selectAll";
+import { priceChange } from "./sumAllPrice";
+import { syncCheckBox } from "./syncCheckBox";
+import { countNumber } from "./updatePrice";
 
 const foodTypeNav = getNode(".food-type__container");
 const AccordCold = getNode("#foodTypeCold");
@@ -51,11 +57,11 @@ async function getCardAddedProductsNew() {
     >
     <div class="food-type__accordion__count price_counter">
       <button type="button" class="minus-button">
-        <img src="/images/minus-button-black.svg" alt="" />
+        <img src="/images/minus-button-black.svg" alt="수량 감소 버튼" />
       </button>
       <span class="counter">${quantity}</span>
       <button type="button" class="plus-button">
-        <img src="/images/plus-button-black.svg" alt="" />
+        <img src="/images/plus-button-black.svg" alt="수량 추가 버튼" />
       </button>
     </div>
 
@@ -78,8 +84,8 @@ async function getCardAddedProductsNew() {
         insertAfter(AccordRoomTemp, templete);
       }
     }
-    new AgreementManager("#selectAll", "accordion__input");
-    new AgreementManager("#selectAll2", "accordion__input");
+    new selectAll("#selectAll", "accordion__input");
+    new selectAll("#selectAll2", "accordion__input");
     countNumber();
     deleteSelected();
     deleteItem();
@@ -90,9 +96,7 @@ async function getCardAddedProductsNew() {
 
 document.addEventListener("DOMContentLoaded", getCardAddedProductsNew);
 
-// 로그인 됐을때 로그인 버튼에서 => 주문하기로 변경
-// 로그인 됐을때 배송지 안보임 => 고객 배송지 정보 받아오기
-// isAuth = false
+// 로그인 상태에 따른 UI 변경
 async function isLogin() {
   const auth = await getStorage("auth");
   const adress = getNode(".adress__client-adress");
@@ -111,280 +115,14 @@ async function isLogin() {
 }
 
 document.addEventListener("DOMContentLoaded", isLogin);
-
-// 체크박스 전체선택 관련 함수
-// 전체선택 관련 함수
-class AgreementManager {
-  constructor(checkAllSelector, checkItemSelector) {
-    this.checkAllElement = document.querySelector(checkAllSelector);
-    this.checkItemElements = document.getElementsByName(checkItemSelector);
-
-    if (this.checkAllElement && this.checkItemElements.length > 0) {
-      this.init();
-    } else {
-      console.warn(
-        "AgreementManager: Required elements are not found on the page."
-      );
-    }
-  }
-
-  init() {
-    this.checkAllElement.addEventListener("change", () => this.checkAll());
-    Array.from(this.checkItemElements).forEach((checkItem) => {
-      checkItem.addEventListener("change", () => this.updateCheckAll());
-    });
-  }
-
-  checkAll() {
-    Array.from(this.checkItemElements).forEach((checkItem) => {
-      checkItem.checked = this.checkAllElement.checked;
-    });
-  }
-
-  updateCheckAll() {
-    const allChecked = Array.from(this.checkItemElements).every(
-      (checkItem) => checkItem.checked
-    );
-    this.checkAllElement.checked = allChecked;
-  }
-}
-
-// // 위아래 체크 박스 동기화
-function syncCheckBox(checkBox) {
-  const checkAllBoxes = document.querySelectorAll(checkBox);
-
-  checkAllBoxes.forEach((box) => {
-    box.addEventListener("change", () => {
-      syncCheckAllBoxes(box.checked);
-    });
-  });
-
-  function syncCheckAllBoxes(isChecked) {
-    checkAllBoxes.forEach((box) => {
-      box.checked = isChecked;
-    });
-  }
-}
 document.addEventListener(
   "DOMContentLoaded",
   syncCheckBox(".checkbox__check-all__box")
 );
-
-// 수량 변경 및 가격변경 함수 (랜더링되고 나서 실행)
-function countNumber() {
-  const counterSets = document.querySelectorAll(".food-type__accordion");
-
-  counterSets.forEach((set) => {
-    const counterElement = set.querySelector(".counter");
-    const minusButton = set.querySelector(".minus-button");
-    const plusButton = set.querySelector(".plus-button");
-    const priceValue = set.querySelector(".food-type__accordion__price__value");
-    const priceDiscount = set.querySelector(
-      ".food-type__accordion__price__discount"
-    );
-    let count = Number(counterElement.textContent);
-    const dataPrice = valueTrimmed(priceValue) / count;
-    const dataDiscount = valueTrimmed(priceDiscount) / count;
-
-    minusButton.addEventListener("click", () => {
-      if (count > 0) {
-        count--;
-        counterElement.textContent = count;
-        priceValue.textContent = `${formatPrice(count * dataPrice)}원`;
-        priceDiscount.textContent = `${formatPrice(count * dataDiscount)}원`;
-      }
-    });
-
-    plusButton.addEventListener("click", () => {
-      count++;
-      counterElement.textContent = count;
-      priceValue.textContent = `${formatPrice(count * dataPrice)}원`;
-      priceDiscount.textContent = `${formatPrice(count * dataDiscount)}원`;
-    });
-  });
-}
 
 // 받은 가격 원단위, 숫자단위 떼기
 function valueTrimmed(value) {
   return value.textContent
     .substr(0, value.textContent.length - 1)
     .replace(",", "");
-}
-
-// 선택항목 삭제 함수
-function deleteSelected() {
-  const templates = getNodes(".food-type__accordion");
-
-  templates.forEach((item) => {
-    const deleteButton = item.querySelector(".food-type__accordion__delete");
-    deleteButton.addEventListener("click", () => {
-      item.remove();
-    });
-  });
-}
-
-// 전체삭제 삭제 함수
-function deleteItem() {
-  const deleteButton = getNode(".checkbox__delete");
-  const checkboxes = document.getElementsByName("accordion__input");
-  deleteButton.addEventListener("click", () => {
-    Array.from(checkboxes).forEach((checkBox) => {
-      if (checkBox.checked) {
-        const item = checkBox.closest(".food-type__accordion");
-        if (item) {
-          item.remove();
-        }
-      }
-    });
-  });
-}
-
-// 선택수량 업데이트 함수
-function updateSelectedCount() {
-  const checkboxes = Array.from(getNodes(".food-type__accordion__input"));
-  const checked = getNode(".checkbox__check-all__label");
-
-  const selectedCount = checkboxes.filter(
-    (checkbox) => checkbox.checked
-  ).length;
-  checked.textContent = `전체선택 (${selectedCount}/${checkboxes.length})`;
-}
-
-// 전체 상품 갯수 랜더링 및 상품 선택시 선택된 상품 갯수 업데이트
-function countChange() {
-  const templates = getNodes(".food-type__accordion");
-  const checkboxes = Array.from(getNodes(".food-type__accordion__input"));
-  const checked = getNode(".checkbox__check-all__label");
-  checked.textContent = `전체선택 (0/${checkboxes.length})`;
-  Array.from(templates).forEach((checkbox) => {
-    const inputBox = checkbox.querySelector(".food-type__accordion__input");
-    const deleteButton = checkbox.querySelector(
-      ".food-type__accordion__delete"
-    );
-    const checkAllInput = document.querySelector("#selectAll");
-    const deleteAllButton = getNode(".checkbox__delete");
-    inputBox.addEventListener("change", updateSelectedCount);
-    deleteButton.addEventListener("click", updateSelectedCount);
-    checkAllInput.addEventListener("change", updateSelectedCount);
-    deleteAllButton.addEventListener("click", updateSelectedCount);
-  });
-}
-
-// 수량 변경시 총액 변경
-
-function updateSumAllPrice() {
-  const templates = getNodes(".food-type__accordion");
-  // 원가
-  const priceArray = Array.from(templates)
-    .filter((item) => {
-      const checkBox = item.querySelector(".food-type__accordion__input");
-      return checkBox.checked;
-    })
-    .map((item) => {
-      const price = item.querySelector(".food-type__accordion__price__value");
-      return Number(valueTrimmed(price));
-    });
-
-  const valueResult = Array.from(priceArray).reduce(function (
-    accumulator,
-    currentValue
-  ) {
-    return accumulator + currentValue;
-  },
-  0);
-
-  // 할인값
-  const discountArray = Array.from(templates)
-    .filter((item) => {
-      const checkBox = item.querySelector(".food-type__accordion__input");
-      return checkBox.checked;
-    })
-    .map((item) => {
-      const price = item.querySelector(
-        ".food-type__accordion__price__discount"
-      );
-      return Number(valueTrimmed(price));
-    });
-
-  const discountPrice = Array.from(discountArray).reduce(function (
-    accumulator,
-    currentValue
-  ) {
-    return accumulator + currentValue;
-  },
-  0);
-
-  const discountResult = valueResult - discountPrice;
-
-  if (valueResult == 0) {
-    document.querySelector(".price__group__delivery").textContent = `0원`;
-  }
-  // 상품금액
-  document.querySelector(".price__group__value").textContent = `${formatPrice(
-    valueResult
-  )}원`;
-  // 상품할인 금액
-  document.querySelector(
-    ".price__group__discounted"
-  ).textContent = `-${formatPrice(discountResult)}원`;
-  // 결제예정 금액
-  document.querySelector(
-    ".total-price__sum__value"
-  ).textContent = `${formatPrice(discountPrice)}`;
-
-  if (0 != discountPrice && discountPrice < 30000) {
-    document.querySelector(
-      ".total-price__sum__value"
-    ).textContent = `${formatPrice(discountPrice + 3000)}`;
-    document.querySelector(
-      ".price__group__delivery"
-    ).textContent = `${formatPrice(3000)}`;
-  }
-}
-
-// function updateSumAllDiscounted() {
-//   const templates = getNodes(".food-type__accordion");
-//   const discountArray = Array.from(templates)
-//     .filter((item) => {
-//       const checkBox = item.querySelector(".food-type__accordion__input");
-//       return checkBox.checked;
-//     })
-//     .map((item) => {
-//       const price = item.querySelector(".food-type__accordion__price__discount");
-//       return Number(valueTrimmed(price));
-//     });
-
-//   const valueResult = Array.from(priceArray).reduce(function (
-//     accumulator,
-//     currentValue
-//   ) {
-//     return accumulator + currentValue;
-//   },
-//   0);
-//   const priceTemplate = document.querySelector(".price__group__di");
-//   document.querySelector(".price__group__value").textContent = `${formatPrice(
-//     valueResult
-//   )}원`;
-// }
-
-// 선택삭제, 선택, 삭제클릭시 updateSumAllPrice 이벤트 추가
-function priceChange() {
-  const templates = getNodes(".food-type__accordion");
-  Array.from(templates).forEach((checkbox) => {
-    const inputBox = checkbox.querySelector(".food-type__accordion__input");
-    const deleteButton = checkbox.querySelector(
-      ".food-type__accordion__delete"
-    );
-    const checkAllInput = document.querySelector("#selectAll");
-    const deleteAllButton = getNode(".checkbox__delete");
-    const minusButton = checkbox.querySelector(".minus-button");
-    const plusButton = checkbox.querySelector(".plus-button");
-
-    inputBox.addEventListener("change", updateSumAllPrice);
-    deleteButton.addEventListener("click", updateSumAllPrice);
-    checkAllInput.addEventListener("change", updateSumAllPrice);
-    deleteAllButton.addEventListener("click", updateSumAllPrice);
-    minusButton.addEventListener("click", updateSumAllPrice);
-    plusButton.addEventListener("click", updateSumAllPrice);
-  });
 }
